src/litecow_server/pyproject.toml
[tool.poetry]
name = "litecow_server"
version = "0.1.0"
description = ""
authors = ["Striveworks"]

[tool.poetry.dependencies]
python = "^3.6"
grpcio = "*"
onnxruntime = "^1.7.0"
numpy = "^1.17.4"
boto3 = "^1.11.14"
pillow = "^8.1.2"
litecow_common = { path = "../litecow_common" }

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
litecow-service = 'litecow_server.service:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
